refactor: apply centralized error handling to cards controller
